pyproject.toml
[tool.poetry]
name = "spurious_emu"
version = "0.4.1"
description = "VBA static and dynamic analysis tool for malware analysts"
authors = ["Louis Dubois <ldbo@protonmail.com>"]
license = "AGPL-3.0-or-later"
readme= "Readme.md"
homepage = "https://github.com/ldbo/SpuriousEmu"
packages = [{include = "emu"}]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Visual Basic",
    "Topic :: Security",
    "Topic :: Software Development :: Interpreters",
]

[tool.poetry.scripts]
emu = "emu.__main__:main"

[tool.poetry.dependencies]
python = "^3.8"
oletools = "^0.55.1"
python-magic = "^0.4.18"
prettytable = "^0.7.2"
pyparsing = "^2.4.7"
lark-parser = "^0.9.0"

[tool.poetry.dev-dependencies]
ipython = "^7.17.0"
nose = "^1.3.7"
coverage = "^5.2.1"
mypy = "^0.782"
networkx = "^2.4"
lxml = "^4.5.2"
matplotlib = "^3.3.1"
ipdb = "^0.13.3"
bump2version = "^1.0.0"
coveralls = "^2.1.2"
black = "^19.10b0"

[tool.black]
line-length = 80
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.venv
    | _build
    | build
    | dist
  )/
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

